pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "pyrise2"
version = "0.0.1"
authors = [
  { name="Jules", email="jules@example.com" },
]
description = "A modern Python package for solving and simulating DSGE models."
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
  "jax[cpu]>=0.4.0",
  "numpy>=1.24.0",
  "scipy>=1.11.0",
  "sympy>=1.12.0",
  "pandas>=2.0.0",
  "pyyaml>=6.0.0",
  "ruamel.yaml>=0.17.0",
  "lark>=1.1.0",
  "matplotlib>=3.7.0"
]

[project.optional-dependencies]
dev = [
    "pytest",
    "black",
    "isort",
    "mypy",
    "flake8",
    "pre-commit"
]

[project.urls]
"Homepage" = "https://github.com/jules/pyRISE2"
"Bug Tracker" = "https://github.com/jules/pyRISE2/issues"
